refactor(capabilities): share confique parse_flag + provider max-in-flight helpers
The per-cap confique migration re-introduced the duplication
Qodana flagged (the original iamacoffeepot/aether#1055 smell):
four identical `parse_flag` bodies (http / gemini / anthropic /
audio) and two identical `parse_max_in_flight` bodies (gemini and
anthropic — same `> 0` filter, same `DEFAULT_MAX_IN_FLIGHT = 2`).

This collapses them into a shared `config_env` module on
`aether-capabilities`:

- `parse_flag` — bool `"1"`/`"true"`, used by all four caps.
- `parse_provider_max_in_flight` — `> 0` filter + soft fall-back
  to `DEFAULT_PROVIDER_MAX_IN_FLIGHT`, shared by gemini and
  anthropic. Each provider's public `DEFAULT_MAX_IN_FLIGHT` now
  aliases the shared const.
- `parse_flag` + `parse_max_in_flight` tests consolidated into
  the shared module's test mod (4 + 2 copies → 2 total).

`parse_timeout_ms` stays per-struct: the defaults differ across
the consumers (http 30 s, gemini 180 s, anthropic 120 s, dag
retention 60 s / 300 s) and confique's literal-only
`#[config(default = …)]` can't be parametrized, so dedup-ing the
ms→Duration parsers across distinct defaults needs a `DurationMs`
newtype — flagged as the per-field-literal-tax follow-up.

129 lines removed, 20 added. Pre-flight green (fmt + clippy +
doc + wasm32 + 1415 nextest).
